# backend/openapi.yaml
openapi: 3.1.0
info:
  title: SNT Portal API
  version: 1.0.0
  description: |
    Unified API for resident/chairman portal.
servers:
  - url: http://localhost:3000
    description: local
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/v1/users:
    get:
      summary: List users (chairman)
      security:
        - bearerAuth: []
    post:
      summary: Create USER with temporary password (chairman)
      security:
        - bearerAuth: []
  /api/v1/plots:
    get:
      summary: List plots
      security:
        - bearerAuth: []
    post:
      summary: Create plot (chairman)
      security:
        - bearerAuth: []
...